pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stag"
version = "0.1.0"
description = "Sensor-based Tracking and Analysis of Gait — unsupervised behaviour classification from wearable accelerometers."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Alexander R. H. Matthews"},
    {name = "Lindsay R. Matthews"},
    {name = "Bart R. H. Geurten", email = "bart.geurten@otago.ac.nz"},
]
keywords = [
    "accelerometer",
    "behaviour-classification",
    "hidden-markov-model",
    "deer",
    "precision-livestock-farming",
    "unsupervised-learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "sqlalchemy>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "tqdm>=4.65",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x",
    # cuml must be installed via conda/rapids — listed here for reference
]
gps = [
    "pyproj>=3.5",
    "tables>=3.8",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=1.23",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3",
]

[project.urls]
Repository = "https://github.com/zerotonin/stag"
Documentation = "https://stag-pipeline.readthedocs.io"

[tool.setuptools.packages.find]
include = ["stag*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "D"]
ignore = ["D100", "D104"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
testpaths = ["tests"]
